LayoutTests/fast/selectors/lang-specificity-expected.txt

Verify selector specifying multiple :lang() pseudo class.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


HTML lang only
PASS document.querySelectorAll(":lang(fr):lang(fr-be):lang(fr) #target1").length is 1
PASS getComputedStyle(document.getElementById("target1")).color is "rgb(1, 2, 3)"
PASS document.querySelectorAll(":lang(fr):lang(fr-ca) #target2").length is 1
PASS getComputedStyle(document.getElementById("target2")).color is "rgb(4, 5, 6)"
PASS document.querySelectorAll(":lang(fr):lang(fr-ch) #target3").length is 1
PASS getComputedStyle(document.getElementById("target3")).color is "rgb(7, 8, 9)"
PASS successfullyParsed is true

TEST COMPLETE
